// src/app/shared/models/common-models/student.model.ts
export interface IStudentModel {
  id?: any;
  firstName?: string;
  lastName?: string;
  middleName?: string;
  dateOfBirth?: Date;
  emailAddress?: string;
  phoneNumber?: string;
  address?: string;
  departmentId?: number;
  guardianId?: number;
  guardianName?: string;
  deptName?: string;
  organizationId?:any;
}